sdm845-common: Rebase on lineage gerrit as of March 2022
Interesting times! So they're finally doing a source-built vendor
bringup in preparation for lineage-19.0 (Android 12) for our
OnePlus SDM845, but testing against lineage-18.1 build tree to
begin with, so some of this was already merged into lineage-19.0
but is applicable to lineage-18.1... but they're trying to freeze
that branch, so it's filed under 19.0.
Let's put a pin in it here, because I think they're about finished
with the Android 11 stuff and are now moving into adapting it for
Android 12 now in the gerrit.

Update README.md & dependencies.