sdm845-common: vibrator: Add vibration interface
Adapted from android_hardware_oneplus, since we aren't building vendor yet.
Trying to find the missing link between having a custom AIDL for the vibrator,
and being able to feed the correct interface to VibrationService.java in AFB.
But this probably gets us closer, and at least it doesn't seem to break.