sdm845-common: init: Add ro.product.model overrides
OnePlus was never very clear about their naming
for different hardware variants with the 6-series,
and we've only gone with forcing A60x3 for
enchilada & fajita instead of differentiating
based on ro.boot props, so it's hard to know if
you have a hardware version with limited LTE bands
(as sold in China & India), compared to the Europe
or North America variants with more LTE bands
available.

Add prop override to set model name based on
rf_version like in later series, so people can know
if they have a China or India market hw variant.